refactor: cms_routes -> cms_router
